Move lock.py into pantsd
lock.py is only impored from in pantsd code. Moving the file
to within pantsd makes the `src/python` directory structure
less cluttered.

[ci skip-rust]

[ci skip-build-wheels]
